requirements.txt
APScheduler==3.8.0
certifi==2023.7.22
charset-normalizer==2.0.6
idna==3.2
pytz==2021.1
PyYAML==6.0.1
redis==3.5.3
requests==2.31.0
six==1.16.0
tzlocal==2.1
urllib3==1.26.18
PyJWT~=2.4.0
cryptography==41.0.6
pip==23.3
wheel==0.38.1
setuptools==65.5.1
future~=0.18.2
newrelic==9.7.0
